refactor(discord): E13 게이트 정리 — 브리지 lease 와이어링 추출 + DeliveryLeaseKey 모듈 분리 + outbound 문서 맵 갱신
turn_bridge/mod.rs 6724→6704 (ceiling 6707), tmux_watcher.rs 7312.
DeliveryLeaseKey → delivery_lease_key.rs (re-export 동일 경로), watcher controller
heartbeat 어댑터 → tmux_watcher/controller_heartbeat.rs (namespace cap).
discord-outbound-migration.md lease guard 섹션 DeliveryLeaseKey 반영. 동작 불변.

Implemented-by: Codex (session 019f2522-19dd-76d0-b8a5-570759b29388)
